flatten top-level stubs
