--- src/http/request.rs
@@ -1,1 +1,0 @@
-pub use crate::sys::http::request::*;
