+++ src/rand.rs
