+++ src/time.rs
@@ -1,6 +1,27 @@
-use super::{Instant, Wait};
-use std::future::IntoFuture;
+//! Async time interfaces.
+
+pub(crate) mod utils {
+    use std::io;
+
+    pub(crate) fn timeout_err(msg: &'static str) -> io::Error {
+        io::Error::new(io::ErrorKind::TimedOut, msg)
+    }
+}
+
+use pin_project_lite::pin_project;
+use std::future::{Future, IntoFuture};
 use std::ops::{Add, AddAssign, Sub, SubAssign};
+use std::pin::Pin;
+use std::task::{Context, Poll};
+
+use crate::{
+    iter::AsyncIterator,
+    runtime::AsyncPollable,
+};
+
+pub use crate::sys::time::SystemTime;
+
+// ---- Duration ----
 
 /// A Duration type to represent a span of time, typically used for system
 /// timeouts.
@@ -161,9 +182,185 @@ impl IntoFuture for Duration {
     }
 }
 
+// ---- Instant ----
+
+/// A measurement of a monotonically nondecreasing clock. Opaque and useful only
+/// with Duration.
+///
+/// This type wraps `std::time::Duration` so we can implement traits on it
+/// without coherence issues, just like if we were implementing this in the
+/// stdlib.
+#[derive(Debug, PartialEq, PartialOrd, Ord, Eq, Hash, Clone, Copy)]
+pub struct Instant(pub(crate) crate::sys::time::MonotonicInstant);
+
+impl Instant {
+    /// Returns an instant corresponding to "now".
+    ///
+    /// # Examples
+    ///
+    /// ```no_run
+    /// use wstd::time::Instant;
+    ///
+    /// let now = Instant::now();
+    /// ```
+    #[must_use]
+    pub fn now() -> Self {
+        Instant(crate::sys::time::now())
+    }
+
+    /// Returns the amount of time elapsed from another instant to this one, or zero duration if
+    /// that instant is later than this one.
+    pub fn duration_since(&self, earlier: Instant) -> Duration {
+        Duration::from_nanos(self.0.saturating_sub(earlier.0))
+    }
+
+    /// Returns the amount of time elapsed since this instant.
+    pub fn elapsed(&self) -> Duration {
+        Instant::now().duration_since(*self)
+    }
+}
+
+impl Add<Duration> for Instant {
+    type Output = Self;
+
+    fn add(self, rhs: Duration) -> Self::Output {
+        Self(self.0 + rhs.0)
+    }
+}
+
+impl AddAssign<Duration> for Instant {
+    fn add_assign(&mut self, rhs: Duration) {
+        *self = Self(self.0 + rhs.0)
+    }
+}
+
+impl Sub<Duration> for Instant {
+    type Output = Self;
+
+    fn sub(self, rhs: Duration) -> Self::Output {
+        Self(self.0 - rhs.0)
+    }
+}
+
+impl SubAssign<Duration> for Instant {
+    fn sub_assign(&mut self, rhs: Duration) {
+        *self = Self(self.0 - rhs.0)
+    }
+}
+
+impl IntoFuture for Instant {
+    type Output = Instant;
+
+    type IntoFuture = Wait;
+
+    fn into_future(self) -> Self::IntoFuture {
+        crate::task::sleep_until(self)
+    }
+}
+
+// ---- Timer / Interval ----
+
+/// An async iterator representing notifications at fixed interval.
+pub fn interval(duration: Duration) -> Interval {
+    Interval { duration }
+}
+
+/// An async iterator representing notifications at fixed interval.
+///
+/// See the [`interval`] function for more.
+#[derive(Debug)]
+pub struct Interval {
+    duration: Duration,
+}
+impl AsyncIterator for Interval {
+    type Item = Instant;
+
+    async fn next(&mut self) -> Option<Self::Item> {
+        Some(Timer::after(self.duration).wait().await)
+    }
+}
+
+#[derive(Debug)]
+pub struct Timer(Option<AsyncPollable>);
+
+impl Timer {
+    pub fn never() -> Timer {
+        Timer(None)
+    }
+    pub fn at(deadline: Instant) -> Timer {
+        let pollable = crate::sys::time::subscribe_at(deadline.0);
+        Timer(Some(pollable))
+    }
+    pub fn after(duration: Duration) -> Timer {
+        let pollable = crate::sys::time::subscribe_after(duration.0);
+        Timer(Some(pollable))
+    }
+    pub fn set_after(&mut self, duration: Duration) {
+        *self = Self::after(duration);
+    }
+    pub fn wait(&self) -> Wait {
+        let wait_for = self.0.as_ref().map(AsyncPollable::wait_for);
+        Wait { wait_for }
+    }
+}
+
+pin_project! {
+    /// Future created by [`Timer::wait`]
+    #[must_use = "futures do nothing unless polled or .awaited"]
+    pub struct Wait {
+        #[pin]
+        wait_for: Option<crate::runtime::WaitFor>
+    }
+}
+
+impl Future for Wait {
+    type Output = Instant;
+
+    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
+        let this = self.project();
+        match this.wait_for.as_pin_mut() {
+            None => Poll::Pending,
+            Some(f) => match f.poll(cx) {
+                Poll::Pending => Poll::Pending,
+                Poll::Ready(()) => Poll::Ready(Instant::now()),
+            },
+        }
+    }
+}
+
 #[cfg(test)]
-mod tests {
+mod test {
     use super::*;
+
+    async fn debug_duration(what: &str, f: impl Future<Output = Instant>) {
+        let start = Instant::now();
+        let now = f.await;
+        let d = now.duration_since(start);
+        let d: std::time::Duration = d.into();
+        println!("{what} awaited for {} s", d.as_secs_f32());
+    }
+
+    #[test]
+    fn timer_now() {
+        crate::runtime::block_on(debug_duration("timer_now", async {
+            Timer::at(Instant::now()).wait().await
+        }));
+    }
+
+    #[test]
+    fn timer_after_100_milliseconds() {
+        crate::runtime::block_on(debug_duration("timer_after_100_milliseconds", async {
+            Timer::after(Duration::from_millis(100)).wait().await
+        }));
+    }
+
+    #[test]
+    fn test_duration_since() {
+        let x = Instant::now();
+        let d = Duration::new(456, 789);
+        let y = x + d;
+        assert_eq!(y.duration_since(x), d);
+    }
 
     #[test]
     fn test_new_from_as() {
